requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
billiard==3.3.0.23
celery==3.1.24
channels==0.17.3
constantly==15.1.0
daphne==0.15.0
Django==1.10.3
djangorestframework==3.5.2
incremental==16.10.1
kombu==3.0.37
msgpack-python==0.4.8
pytz==2016.7
redis==2.10.5
Twisted==16.5.0
txaio==2.5.1
zope.interface==4.3.2
